libfimg: Uninline fimgQueue
Long time ago when fimgQueue was still fimgWrite, it was a simple one
line accessor, for which inlining was fully justified. With the time it
has evolved to fimgQueue and become a bit more complex, while the inline
keyword stayed, increasing code size without any reason and in some
cases even reducing performance. This patch finally fixes this.

Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>